requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
Flask==2.1.1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
lxml==4.8.0
MarkupSafe==2.1.1
numpy==1.22.3
pandas==1.4.2
Pillow==9.1.0
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.1
requests==2.27.1
six==1.16.0
typing==3.7.4.3
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
